feat: parsing over never type
We introduces the never type to represent a type that is never
instantiated. This is useful for cases where a branch of an if statement
never returns for instance:

```mun
let b = if a {
  3
} else {
  return;
}
```

In the above case `b` should be assigned the type `int` because the then
branch has type `int` and the else branch has type `never`.