perf: use a slab allocator for `ObjectInfo`
This should improve performance of the GC by limiting allocations of
`ObjectInfo` to large chunks.

Originally, I was hoping to use the allocator from the `slab` crate out
of the box, but that would cause all of the values to move when ever it
needed to resize. I added the `pinned_slab` crate to get around that
limitation.

closes #106