improvement(runtime): clarify usage of unsafe TypeInfo
Garbage collection uses TypeInfo pointers that cannot directly be
linked to the lifetime of their assemblies. As a result any usage of
these pointers is unsafe. To clarify this, the RawTypeInfo type has been
renamed to UnsafeTypeInfo and everywhere it is used, unsafe must be
added, similar to Rust's UnsafeCell type.