feat: assoc method completion (#578)
This PR refactors how we analyze the context of completions and adds
completion of associated methods.

There is some unused code in this PR which I need for adding more
completions (like methods and use statements). I didnt add them to this
PR yet to make sure it doesnt grow to big.